[refactor] remove imaging/autogamma.py shim, import directly from autofunctions.gamma
