+++ fibsem/imaging/__init__.py
@@ -1,5 +1,5 @@
 from fibsem.imaging.spot import *
-from fibsem.imaging.autogamma import *
+from fibsem.autofunctions.gamma import auto_gamma, apply_gamma, apply_clahe
 from fibsem.imaging.masks import *
 from fibsem.imaging.tiled import *
 from fibsem.imaging.utils import *
