--- fibsem/imaging/autogamma.py
@@ -1,3 +1,0 @@
-from fibsem.autofunctions.gamma import auto_gamma, apply_gamma, apply_clahe
-
-__all__ = ["auto_gamma", "apply_gamma", "apply_clahe"]
